# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is server for Archivit API.",
        "title": "ARCHIVIT API",
        "contact": {},
        "version": "latest"
    },
    "paths": {
...
        "/file/delete": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "delete file",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "delete file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "file txt and m4a",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid JWT",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/file/download/all": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "download all files",
                "produces": [
                    "application/json"
                ],
                "summary": "download all files",
                "responses": {
                    "200": {
                        "description": "All files downloaded successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid JWT",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}